repository: move crypto package to internal/repository/crypto
