+++ internal/repository/crypto/crypto_test.go
@@ -6,7 +6,7 @@ import (
 	"io"
 	"testing"
 
-	"github.com/restic/restic/internal/crypto"
+	"github.com/restic/restic/internal/repository/crypto"
 	rtest "github.com/restic/restic/internal/test"
 
 	"github.com/restic/chunker"
