+++ internal/repository/pack/pack.go
@@ -12,7 +12,7 @@ import (
 	"github.com/restic/restic/internal/errors"
 	"github.com/restic/restic/internal/restic"
 
-	"github.com/restic/restic/internal/crypto"
+	"github.com/restic/restic/internal/repository/crypto"
 )
 
 // ErrBroken is returned by Add and Finalize after a write error. The packer
